# doc/api.yaml
openapi: 3.0.3
servers:
  - url: "http://localhost:3000"
info:
  title: WasaText
  description: |
    Connect with your friends effortlessly using WASAText! Send and receive messages, whether one-on-one
    or in groups, all from the convenience of your PC. Enjoy seamless conversations with text or GIFs and
    easily stay in touch through your private chats or group discussions.
  version: 0.0.1
security:
  - bearerAuth: []
...
paths:
...
  /users:
    parameters:
      - name: found_user
        in: query
        required: true
        description: Username to search
        schema: { $ref: "#/components/schemas/Name" }
    # Search users
    get:
      tags: ["User"]
      operationId: findUsers
      summary: Find users
      description: Find users by name.
      responses:
        "200":
          description: Users found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Users" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/ServerError" }
  /users/{user}/username:
    parameters:
      - $ref: "#/components/parameters/uid"
    put:
      tags: [User]
      operationId: setMyUserName
      summary: Set username
      description: Updates current username
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UsernameInput" }
        required: true
      responses:
        "200":
          description: Successful
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User"}
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /users/{user}/photo:
    parameters:
      - $ref: "#/components/parameters/uid"
    put:
      tags: [User]
      operationId: setMyPhoto
      summary: Set photo
      description: Updates profile picture. Send an empty string to remove the photo.
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PhotoInput" }
        required: true
      responses:
        "200":
          description: The photo was uploaded successfully
          content:
            plain/text:
              schema: { $ref: "#/components/schemas/Success" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }

  # Tag Conversation
  /users/{user}/conversations:
  # List my conversations
    parameters:
      - $ref: "#/components/parameters/uid"
    get:
      tags: [Conversation]
      operationId: getMyConversations
      summary: List my conversations
      description: Get all user conversations
      responses:
        "200":
          description: Successfully retrieved conversations
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConversationCollection" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/ServerError" }
    
    # Create a new conversation
...
components:
...
  parameters:
    uid:
      name: user
      in: path
      required: true
      description: User ID
      schema: { $ref: "#/components/schemas/ResourceId" }
...
  schemas:
    Success:
      title: Successful
      description: The operation was successful
      type: string
      pattern: "^.*?$"
      example: "Action successfully completed"
      minLength: 2
      maxLength: 30
    # Elements
    ResourceId:
      title: ID - identification
      description: General identification number (used for example for users, groups or conversation)
      type: integer
      example: 100
      minimum: 1
    Photo:
      title: Photo
      description: Photo of a user profile or a group
      type: string
      format: byte
      minLength: 0
      maxLength: 9999
    Name:
      title: Name
      description: Name of a user or a group
      type: string
      pattern: ^[a-zA-Z0-9 _-]{3,30}$
      example: "user123"
      minLength: 3
      maxLength: 13
    Timestamp:
      title: Timestamp
      description: The timestamp of the post.
      type: string
      format: date-time
      example: "2020-12-31T23:59:59"
      minLength: 19
      maxLength: 19
    User:
      title: User profile
      description: User profile information
      type: object
      properties:
        username: { $ref: "#/components/schemas/Name" }
        photo: { $ref: "#/components/schemas/Photo" }
        user_id: { $ref: "#/components/schemas/ResourceId" }
      required: [username, user_id]
      example:
        username: "MarioRossi"
        user_id: 10
    Users:
      title: List of profile
      description: List of users in a group, conversation or in the search operation
      type: object
      properties:
        users:
          type: array
          description: All users returned after a request like search
          items: { $ref: "#/components/schemas/User" }
          uniqueItems: true
          minItems: 0
          maxItems: 100
      required: [users]
      example:
        users:
          - username: "MarioRossi"
            user_id: 10
    Text:
      title: Text
      description: Used to represent a text in a message or comment
      type: string
      example: "Hi!"
      minLength: 1
      maxLength: 100
      pattern: "^.*?$"
    ContentText:
      type: object
      properties:
        text: { $ref: "#/components/schemas/Text" }
      required: [text]
    ContentPhoto:
      type: object
      properties:
        photo: { $ref: "#/components/schemas/Photo" }
        caption: { $ref: "#/components/schemas/Text" }
      required: [photo]
    BodyMessage:
      description: The content of the message (text or photo)
      type: object
      oneOf:
        - $ref: "#/components/schemas/ContentText"
        - $ref: "#/components/schemas/ContentPhoto"
...
    Message:
      type: object
      description: A message in a conversation
      properties:
        message_id: { $ref: "#/components/schemas/ResourceId" }
        conversation_id: { $ref: "#/components/schemas/ResourceId" }
        body: { $ref: "#/components/schemas/BodyMessage" }
        sender_id: { $ref: "#/components/schemas/ResourceId" } # Chi lo ha mandato
        sender_name: { $ref: "#/components/schemas/Name" } # Opzionale se sono io
        send_time: { $ref: "#/components/schemas/Timestamp" }
        reply_to:
          allOf: # Usa allOf per combinare il riferimento con la descrizione
            - $ref: "#/components/schemas/ResourceId"
          description: "If present, this message is a reply to the message with this ID"
        status:
          type: string
          description: Status of the message delivery
          enum: [sent, received, read]
      required: [message_id, body, send_time]
      example:
        message_id: 123
        conversation_id: 456
        body:
          text: "Hello!"
        sender_id: 10
        send_time: "2023-01-01T12:00:00"
        status: "sent"
...
    ConversationSummary:
      title: Conversation
      description: Conversation is in a group or between two users
      type: object
      properties:
        conversation_id: { $ref: "#/components/schemas/ResourceId" }
        name: { $ref: "#/components/schemas/Name" }
        photo: { $ref: "#/components/schemas/Photo" }
        last_msg: { $ref: "#/components/schemas/Message" }
    ConversationCollection:
      title: List of conversation
      description: All conversation of a user
      type: object
      properties:
        conversations:
          type: array
          description: Contain all conversation of the user
          items: { $ref: "#/components/schemas/ConversationSummary" }
          minItems: 0
          maxItems: 100
      required:
        - conversations
      example:
        conversations: []
    # Input Schemas for Request Bodies
...
    UsernameInput:
      description: New name to replace the current user name
      type: object
      properties:
        name: { $ref: "#/components/schemas/Name" }
      required:
        - name
    PhotoInput:
      description: Photo to upload in the profile or group
      type: object
      properties:
        photo: { $ref: "#/components/schemas/Photo" }
      required:
        - photo
...
    ErrorItem:
      type: object
      description: Description of the error
      properties:
        code:
          description: Code error
          type: string
          example: "not_found"
          enum: ["bad_request", "unauthorized", "forbidden", "not_found", "conflict", "payload_too_large", "server_error"]
        reason:
          type: string
          description: Explanation of the reason for the error which can be shown to a client user
          pattern: "^.*?$"
          example: Page not found
          minLength: 1
          maxLength: 100
        message:
          type: string
          description: More details of the error
          pattern: "^.*?$"
          example: "The requested URL was not found"
          minLength: 1
          maxLength: 100
        more_info:
          type: string
          description: URL or text with more info
      required:
        - code
        - reason
    Error:
      description: Object describe an error with the code and the message
      type: object
      properties:
        trace:
          type: string
          description: Error trace id
        errors:
          description: Array of error codes generated by the operation
          type: array
          items: { $ref: "#/components/schemas/ErrorItem" }
          minItems: 1
          maxItems: 100
      required:
        - errors
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    BadRequest:
      description: The request was not compliant with the documentation (eg. missing fields, etc)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: The access token is missing or it's expired
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: The user is not allowed to perform the action
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Element not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ServerError:
      description: The server encountered an internal error. Further info in server logs
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: The resource already exists (e.g. username already taken)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    PayloadTooLarge:
      description: The request entity is larger than limits defined by server
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }